Reuse module imports for field name bindings
