+++ tests/data/expected/main/jsonschema/field_name_bindings/typing/MsgspecStruct_3.10_False.py
@@ -1,0 +1,17 @@
+# generated by datamodel-codegen:
+#   filename:  typing.json
+
+from __future__ import annotations
+
+from typing import Annotated as Annotated_aliased
+from typing import Union as Union_aliased
+
+from msgspec import UNSET, Meta, Struct, UnsetType
+
+
+class Record(Struct):
+    Union: Union_aliased[str, UnsetType] = 'list'
+    Annotated: Union_aliased[str, UnsetType] = 'list'
+    choice: Union_aliased[int, str, UnsetType] = UNSET
+    count: Union_aliased[Annotated_aliased[int, Meta(ge=0)], UnsetType] = UNSET
+    list: Union_aliased[str, UnsetType] = UNSET
