+++ tests/data/expected/main/jsonschema/field_name_bindings/typing/PydanticV2BaseModel_3.10_False.py
@@ -1,0 +1,18 @@
+# generated by datamodel-codegen:
+#   filename:  typing.json
+
+from __future__ import annotations
+
+from typing import Annotated as Annotated_aliased
+from typing import Optional
+from typing import Union as Union_aliased
+
+from pydantic import BaseModel, Field
+
+
+class Record(BaseModel):
+    Union: Optional[str] = 'list'
+    Annotated: Optional[str] = 'list'
+    choice: Optional[Union_aliased[int, str]] = None
+    count: Annotated_aliased[Optional[int], Field(ge=0)] = None
+    list: Optional[str] = None
